requirements.txt
# Web Framework
flask==3.0.2
streamlit==1.32.0

# Image Processing
opencv-python==4.9.0.80
pillow==10.2.0
pdf2image==1.17.0

# OCR & NLP
pytesseract==0.3.10
spacy==3.7.4
nltk==3.8.1

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.4.1.post1

# Database
sqlalchemy==2.0.28
alembic==1.13.1

# Utils
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4
pandas==2.2.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 